File classes refactoring
