+++ WebApp/src/uk/ac/exeter/QuinCe/data/Export/ExportConfig.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QuinCe.data;
+package uk.ac.exeter.QuinCe.data.Export;
 
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
