+++ WebApp/src/uk/ac/exeter/QuinCe/data/Export/ExportException.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QuinCe.data;
+package uk.ac.exeter.QuinCe.data.Export;
 
 /**
  * Exception class for errors encountered in data export mechanism
