+++ WebApp/src/uk/ac/exeter/QuinCe/data/Files/FileExistsException.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QuinCe.database.files;
+package uk.ac.exeter.QuinCe.data.Files;
 
 import javax.sql.DataSource;
