+++ WebApp/src/uk/ac/exeter/QuinCe/data/Files/FileStoreException.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QuinCe.database.files;
+package uk.ac.exeter.QuinCe.data.Files;
 
 /**
  * An exception class for errors in the on-disk data file store
