+++ WebApp/src/uk/ac/exeter/QuinCe/data/Files/RawDataFileException.java
@@ -1,4 +1,4 @@
-package uk.ac.exeter.QuinCe.data;
+package uk.ac.exeter.QuinCe.data.Files;
 
 import java.util.Calendar;
