+++ WebApp/src/uk/ac/exeter/QuinCe/database/Calculation/RawDataDB.java
@@ -20,9 +20,9 @@ import org.joda.time.DateTimeZone;
 import uk.ac.exeter.QCRoutines.config.RoutinesConfig;
 import uk.ac.exeter.QCRoutines.messages.Flag;
 import uk.ac.exeter.QuinCe.data.DateTimeParseException;
-import uk.ac.exeter.QuinCe.data.RawDataValues;
 import uk.ac.exeter.QuinCe.data.Calculation.GasStandardMean;
 import uk.ac.exeter.QuinCe.data.Calculation.GasStandardRuns;
+import uk.ac.exeter.QuinCe.data.Files.RawDataValues;
 import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.data.Instrument.InstrumentException;
 import uk.ac.exeter.QuinCe.data.Instrument.Standards.GasStandardDB;
