+++ WebApp/src/uk/ac/exeter/QuinCe/web/files/FileListBean.java
@@ -11,17 +11,17 @@ import javax.faces.context.ExternalContext;
 import javax.faces.context.FacesContext;
 import javax.sql.DataSource;
 
-import uk.ac.exeter.QuinCe.data.ExportConfig;
-import uk.ac.exeter.QuinCe.data.ExportException;
-import uk.ac.exeter.QuinCe.data.ExportOption;
-import uk.ac.exeter.QuinCe.data.FileInfo;
+import uk.ac.exeter.QuinCe.data.Export.ExportConfig;
+import uk.ac.exeter.QuinCe.data.Export.ExportException;
+import uk.ac.exeter.QuinCe.data.Export.ExportOption;
+import uk.ac.exeter.QuinCe.data.Files.DataFileDB;
+import uk.ac.exeter.QuinCe.data.Files.FileDataInterrogator;
+import uk.ac.exeter.QuinCe.data.Files.FileInfo;
 import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.data.Instrument.InstrumentDB;
 import uk.ac.exeter.QuinCe.database.DatabaseException;
 import uk.ac.exeter.QuinCe.database.DatabaseUtils;
 import uk.ac.exeter.QuinCe.database.RecordNotFoundException;
-import uk.ac.exeter.QuinCe.database.files.DataFileDB;
-import uk.ac.exeter.QuinCe.database.files.FileDataInterrogator;
 import uk.ac.exeter.QuinCe.jobs.JobManager;
 import uk.ac.exeter.QuinCe.jobs.files.FileJob;
 import uk.ac.exeter.QuinCe.utils.MissingParamException;
